Created a wrapper to handle ctrl-c on windows
The only way to interrupt ACL2 is to send a CTRL-c, but java can't
get the process ID and can't send a CTRl-c by itself anyway. As a
way to try to deal with that, I made a c++ program that runs the
ACL2 executable and pipes stdin and stdout, but translates 0 bytes
on stdin into CTRL-Cs. It is a mess right now, and doesn't handle
shutdown well. But it works a little.