Forms named def* autoindent like defun
I was writing defrandoms and proof pad was indenting them like

(defrandom foo ()
           (something here),

which is awkward. I changed the indentation rule to treat any
symbol starting with "def" like it treats defun, only indenting
a little. If I remember right, this is how emacs behaves, so it
may not be completely crazy.