Added namespacing capability.
We need a way to avoid name collisions with user code. Packages are
one way we could do that, but that seems to come with quite a bit of
overhead when you want to certify things, and as far as I can tell
the end-user would have to actually do something to use things from
the package (like PACKAGE::WHATEVER).

Namespaces are simpler. It just mangles names beginning with a
percent sign, prepending the current package name. The usage goes

(push-namespace example)

(defun% %private-thing (x)
  (+ x 1))           ; admits EXAMPLE%PRIVATE-THING

(defun% public-thing (x)
  (/ (%private-thing x) 2)) ; admits PUBLIC-THING

(pop-namespace example)

A user of our book doesn't even need to know what we used a namespace.