Random strings and symbols are generated faster
DrAcula chooses lengths of strings and symbols using a Bernoulli
distribution, so defrandom was as well. It was doing it by actually
simulating a bunch of independent trials and counting how many ended up
positive, which is how the distribution is actually defined, but that was
not very efficient. Now, it has hard-coded weights for the different
results and chooses from them, so it has to generate only one random
number instead of many.